restructure tabby_common::index
